worker/pyproject.toml
[tool.poetry]
name = "docuverse-worker"
version = "1.0.0"
description = "DocuVerse Worker - Async Document Processing Service"
authors = ["DocuVerse Team <team@docuverse.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.5.3"
pymongo = "^4.13.2"
python-dotenv = "^1.1.1"
filetype = "1.2.0"
flower = "2.0.1"
llama-index = "*"
openai = "*"
pinecone = "*"
pandas = "*"
python-docx = "*"
openpyxl = "*"
pytesseract = "*"
pillow = "*"
pdfplumber = "*"
pymupdf = "*"
redis = "^5.0.0"
llama-index-vector-stores-pinecone = "^0.6.0"
llmsherpa = "*"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-celery = "^0.0.0"

# Formatting and Linting
black = "^24.0.0"
isort = "^5.12.0"
ruff = "^0.4.0"
mypy = "^1.13.0"

# Type checking
types-requests = "^2.32.0"
types-PyYAML = "^6.0.12"
types-aiofiles = "^23.2.0"
types-redis = "^4.6.0"

# Development tools
pre-commit = "^4.0.0"
ipython = "^8.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]
known_third_party = ["celery", "redis", "pymongo", "llama_index", "openai", "pinecone", "pandas", "python_docx", "openpyxl", "pytesseract", "pillow", "pdfplumber", "pymupdf"]

# Ruff configuration (modern, fast linter)
[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "ARG001", # unused-function-args
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "Q",  # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT",  # flake8-pytest-style
    "LOG", # flake8-logging-format
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PD",  # pandas-vet
    "PGH", # pygrep-hooks
    "PL",  # pylint
    "TRY", # tryceratops
    "NPY", # numpy-vet
    "AIR", # flake8-airflow
    "PERF", # perflint
    "FURB", # refurb
    "C90", # mccabe
    "ICN", # isort
    "G",   # flake8-logging-format
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "TID", # flake8-tidy-imports
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "A",   # flake8-builtins
    "COM", # flake8-commas
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "ISC", # flake8-implicit-str-concat
    "G",   # flake8-logging-format
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLF", # flake8-self
    "SLOT", # flake8-slots
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "UP",  # pyupgrade
    "YTT", # flake8-2020
    "S",   # bandit
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "PLR0913", # too many arguments
    "PLR0915", # too many statements
    "PLR0912", # too many branches
    "PLR0911", # too many return statements
    "PLR2004", # magic value used in comparison
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["S101", "PLR2004", "PLR0913", "PLR0915"]

# Mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true
color_output = true
error_summary = true

# Exclude patterns
exclude = [
    "build",
    "dist",
    "node_modules",
    ".venv",
    "venv",
    "env",
    "__pycache__",
    "*.pyc",
    "*.pyo",
    "*.pyd",
    ".pytest_cache",
    ".mypy_cache",
    ".coverage",
    "htmlcov",
    "tests",
]

# Per-module options
[[tool.mypy.overrides]]
module = [
    "llama_index.*",
    "pinecone.*",
    "pymongo.*",
    "redis.*",
    "celery.*",
    "flower.*",
    "pytesseract.*",
    "PIL.*",
    "pdfplumber.*",
    "pymupdf.*",
    "openpyxl.*",
    "python_docx.*",
    "pandas.*",
    "numpy.*",
    "requests.*",
    "beautifulsoup4.*",
    "striprtf.*",
    "aiofiles.*",
    "python_jose.*",
    "passlib.*",
    "websockets.*",
    "uvicorn.*",
    "fastapi.*",
    "pydantic.*",
    "pydantic_settings.*",
    "python_multipart.*",
    "python_dotenv.*",
    "tqdm.*",
    "fsspec.*",
    "pypdf.*",
    "pypdf2.*",
    "llama_cloud.*",
    "llama_cloud_services.*",
    "llama_parse.*",
    "llama_index_core.*",
    "llama_index_readers_file.*",
    "llama_index_llms_openai.*",
    "llama_index_agent_openai.*",
    "llama_index_program_openai.*",
    "llama_index_question_gen_openai.*",
    "llama_index_multi_modal_llms_openai.*",
    "llama_index_indices_managed_llama_cloud.*",
    "llama_index_cli.*",
    "llama_index_instrumentation.*",
    "llama_index_embeddings_openai.*",
    "llama_index_vector_stores_pinecone.*",
]
ignore_missing_imports = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=app --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]


